+++ code/packages/kotlin/canonical-cbor/src/main/kotlin/com/codingadventures/canonicalcbor/CborException.kt
@@ -22,4 +22,3 @@ class CborException(val id: String) : Exception(messageFor(id)) {
         }
     }
 }
-